Widened job output class to LongWritable

+++ src/main/java/com/griddynamics/bigdata/CountingReducer.java
@@ -1,6 +1,6 @@
 package com.griddynamics.bigdata;
 
-import org.apache.hadoop.io.IntWritable;
+import org.apache.hadoop.io.LongWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.Reducer;
 
@@ -10,11 +10,11 @@ import java.util.stream.StreamSupport;
 /**
  * Created by msigida on 11/24/15.
  */
-public class CountingReducer extends Reducer<Text, IntWritable, Text, IntWritable> {
+public class CountingReducer extends Reducer<Text, LongWritable, Text, LongWritable> {
 
     @Override
-    protected void reduce(Text key, Iterable<IntWritable> values, Context context) throws IOException, InterruptedException {
-        int sum = StreamSupport.stream(values.spliterator(), false).mapToInt(IntWritable::get).sum();
-        context.write(key, new IntWritable(sum));
+    protected void reduce(Text key, Iterable<LongWritable> values, Context context) throws IOException, InterruptedException {
+        long sum = StreamSupport.stream(values.spliterator(), false).mapToLong(LongWritable::get).sum();
+        context.write(key, new LongWritable(sum));
     }
 }
