+++ src/main/java/com/griddynamics/bigdata/ua/UAMapper.java
@@ -3,7 +3,6 @@ package com.griddynamics.bigdata.ua;
 import com.griddynamics.bigdata.xml.XMLMapper;
 import net.sf.uadetector.UserAgentStringParser;
 import net.sf.uadetector.service.UADetectorServiceFactory;
-import org.apache.hadoop.io.IntWritable;
 import org.apache.hadoop.io.LongWritable;
 import org.apache.hadoop.io.Text;
 import org.w3c.dom.Document;
@@ -14,7 +13,7 @@ import java.io.IOException;
 /**
  * Created by msigida on 11/24/15.
  */
-public class UAMapper extends XMLMapper<LongWritable, Text, IntWritable> {
+public class UAMapper extends XMLMapper<LongWritable, Text, LongWritable> {
 
     private UserAgentStringParser parser;
